requirements.txt
Flask==2.3.3
mongoengine==0.27.0
Flask-WTF==1.1.1
Flask-Login==0.6.2
email-validator==2.0.0
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
WTForms==3.0.1
pymongo==4.6.3
blinker==1.6.2
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
dnspython==2.6.1
